Rename methods

+++ test/TorchSharpTest/TestDataLoader.cs
@@ -12,7 +12,7 @@ namespace TorchSharp
         private class TestDataset : torch.utils.data.Dataset
         {
             public override long Count { get; } = 10;
-            public override Dictionary<string, torch.Tensor> GetTensor(int index)
+            public override Dictionary<string, torch.Tensor> GetTensor(long index)
             {
                 return new() {{"data", torch.tensor(1)}, {"label", torch.tensor(13)}, {"index", torch.tensor(index)}};
             }
